Add files via upload
For those of you with 'dumb' washing machines, that either don't beep when finished, or the beep is ineffective!!
This is MY extended implementation of the excellent work done by Mat at NotEnoughTech.com. Props to him.
https://notenoughtech.com/home-automation/washing-machine-notifications/

This flow monitors the washing machine current profile, and tracks its progress through a wash.
If it detects a wash cycle has been completed (an average is attained over 5 data captures), it will send a notification (I have google home speakers). It will continue monitoring, and if it doesn't detect a power off, it will consistently NAG you to empty it!

Important Configurations to note:
Clearly, you will need a power monitoring device inline (Local Bytes do a smart plug, already flashed with tasmota), and for reliable and relatively quick readings, set the logging telemetry period to 60 seconds. You will need to correspondingly set this figure in the 'Settings' code too:
* flow.set("MetricFrequency", 60);  // how often washing machine reports back
* var standbyCurrent = {
    "LowLimit": 0.04,
    "HighLimit": 0.085
}; - Please observe your washing machine's current trends. This array must be set accordingly to detect a standby state.

Dependencies:
* ntfy (or whatever you have for your notifications)
* Google home speaker integration

Help:
* The 'Context'->'Flow' screen is invaluable for debugging here.